hiding stream api key and moving Channel View to Compose
