+++ android/src/main/AndroidManifest.xml
@@ -26,11 +26,6 @@
             </intent-filter>
         </activity>
 
-        <activity
-            android:name=".android.ChannelActivity"
-            android:windowSoftInputMode="adjustResize"
-            />
-
         <receiver android:name="co.touchlab.droidcon.service.NotificationPublisher" />
 
         <receiver
