--- android/src/main/java/co/touchlab/droidcon/android/ChannelActivity.kt
@@ -1,44 +1,0 @@
-package co.touchlab.droidcon.android
-
-import android.content.Context
-import android.content.Intent
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import io.getstream.chat.android.compose.ui.messages.MessagesScreen
-import io.getstream.chat.android.compose.ui.theme.ChatTheme
-import io.getstream.chat.android.compose.viewmodel.messages.MessagesViewModelFactory
-
-class ChannelActivity : ComponentActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        // 1 - Load the ID of the selected channel
-        val channelId = intent.getStringExtra(KEY_CHANNEL_ID)!!
-
-        // 2 - Add the MessagesScreen to your UI
-        setContent {
-            ChatTheme {
-                MessagesScreen(
-                    viewModelFactory = MessagesViewModelFactory(
-                        context = this,
-                        channelId = channelId,
-                        messageLimit = 30
-                    ),
-                    onBackPressed = { finish() }
-                )
-            }
-        }
-    }
-
-    // 3 - Create an intent to start this Activity, with a given channelId
-    companion object {
-        private const val KEY_CHANNEL_ID = "channelId"
-
-        fun getIntent(context: Context, channelId: String): Intent {
-            return Intent(context, ChannelActivity::class.java).apply {
-                putExtra(KEY_CHANNEL_ID, channelId)
-            }
-        }
-    }
-}
